use gate to check roles and permission

+++ app/Http/Controllers/Admin/UserController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\User;
 use App\Role;
+use Gate;
 use Illuminate\Http\Request;
 
 class UserController extends Controller
@@ -32,6 +33,10 @@ class UserController extends Controller
      */
     public function edit(User $user,Role $role)
     {
+        if(Gate::denies('edit-user')){
+            return redirect(route('admin.users.index'));
+        }
+
         $roles = $role->all();
         return view('admin.users.edit')->with([
             'roles' => $roles,
@@ -60,6 +65,9 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        if(Gate::denies('delete-user')){
+            return redirect(route('admin.users.index'));
+        }
         $user->roles()->detach();
         $user->delete();
